Set modification time in past to ensure editor file write detection
When the editor writes to a file quickly, the filesystem's timestamp
granularity may not be sufficient to detect the change. By setting the
modification time to a point in the past (1 day before current time),
we ensure that any write operation by the editor will result in a
newer timestamp, making it possible to reliably detect when the editor
has modified the file.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>