Fix parsing of filenames that start with a number
The previous implementation incorrectly assumed that all lines
starting with digits were in the format "number:content". This
triggered asserts when filenames themselves started with digits, as
the parsing logic would try to interpret the filename as a line
number.

Parse the rg output only once and use only the empty line separator to
identify the next filename.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>