Change keybinding from C-c C-x g to C-c r
The previous keybinding C-c C-x g was inconvenient because the C-c C-x
prefix is unavailable when editing bash scripts, making it difficult
to use rg-edit-git in that context. The new binding C-c r is more
accessible and should avoid conflicts with other modes while being
quicker to type.

The rg-edit-git-conflicts command can still be invoked with a keybinding
using the C-u prefix, so C-u C-c r will invoke it by default.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>