Check gptel-version to enable the gptel workaround only if needed
The fix for the gptel-rewrite-directives-hook upstream bug was
committed as 129032fca88f29c20343e973c98fa17db3000405 and the
gptel-version variable was introduced in the following commit
f4344b8a7950fd6b969b32f84f0fe427a9bc925b. So now it's possible to
check for the existence of gptel-version to determine whether the
workaround is needed.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>