Enforce permissions on temporary files and directories
The tempfile crate defaults to 0o600 permissions on temporary files,
but unexpectedly it does not enforce 0o700 permissions on temporary
directories created by tempdir().

So this change explicitly sets permissions for both the temporary
directory and the temporary files to stop depending on the default
behavior of the tempfile crate that is bad in the tempdir case.

The change also cleanups the file creation logic.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>